Fix bazaar revisions with basic logging/revision parsing.
The basic case for revision parsing, with no arguments was selecting 'submit:'
for the base revision. This works, but it references the revision number in the
parent branch, rather than the revision number in the current branch. This
works for diffs but 'bzr log' can't log across branches.

Instead of 'submit:', I've changed it to use 'parent:'. This should be a safe
alternative, and in my tests, returns the same revno, just in the current
branch instead of the parent.

Testing done: Ran unit tests