Use parsed revision for SVN diffs.
This is a relatively simple refactor to go through the revision parsing scheme
for the various different svn diffs. The only difference that should be visible
is that it's now possible to pass a changelist name to --revision-range.

Testing done: Ran unit tests

Reviewed at https://reviews.reviewboard.org/r/5164/