Add support for posting with Markdown-formatted commit messages.
This adds a new option `--markdown` option to `rbt post` which prevents
escaping of content within the commit message for the description. This
prevents the need to unescape the description before publishing if the
commit message is considered Markdown-safe, or if it's otherwise
Markdown-formatted.

There's also a `MARKDOWN` option for `.reviewboardrc` that can be set to
enable this by default.

This only works for Review Board 2.0+. For now, it may cause an error
when using it with older versions. We'll likely want to introduce a
capability and make use of it.