Calculate outgoing Mercurial changesets from current revision
Previously, RBTools calculated outgoing Mercurial changesets by asking
Mercurial about *all* changesets on the current branch not in the
remote. While this works, it's not ideal. This querying is expensive
since it pulls in all changesets, not just ancestors of the eventual
changesets.

With this patch, we now pass "-r ." into hg outgoing. This limits the
lookup to changesets that are ancestors of the working copy. In some
repository configurations (such as my own, where I have bookmarks
tracking thousands of changesets that will never be pushed to the
remote), this drastically speeds up diff computation and makes RBTools
faster.

Reviewed at https://reviews.reviewboard.org/r/5066/