Refactor extract_{summary,description} to use parsed revisions.
The methods to extract the summary and description (used for guessing the
fields for posting, and for matching open review requests with -u) were using
the old revision range, and were a bit ad-hoc. This cleans them up to always
use the parsed revision spec (except in the case of hgsubversion, which is a
pretty "special" case).

Testing done:
* Ran unit tests
* Posted this change with -g using this code.

Reviewed at https://reviews.reviewboard.org/r/5221/