Optimize checking vob UUIDs with several ClearCase repositories.
When the server has several configured ClearCase repositories, we would make
HTTP calls for each one (using _get_repository_info()). This could potentially
add a lot of overhead.

This change builds an ordered list of ClearCase repositories, with the ones
that have a matching vobstag at the front. This list is then checked in order
to match the UUIDs, which makes it much faster.

This also has a fix for a side issue that I caught with restricted
repositories. In this case, the request would return with HTTP 403, which would
abort.

Testing Done:
- Posting reviews against authorized clearcase repo with matching vobstag still
  works.
- Posting reviews against authorized clearcase repo with no matching vobstag
  will work, even when there are some unauthorized clearcase repos in the
  list.
- Posting reviews against unauthorized clearcase repo returns an error.

Bugs closed: 3177

Reviewed at https://reviews.reviewboard.org/r/5166/