Split up --repository-url and REPOSITORY into new options/settings.
--repository-url and REPOSITORY had both previously allowed either a
path to be specified, or a repository name. We'd try for the name when
looking up a repository, but would try for the path in different ways
within the SCMClients. In the case of Subversion, this broke pretty
badly when using a repository name.

Now --repository-url has been split into --repository-url and
--repository, while REPOSITORY has been split into REPOSITORY_URL and
REPOSITORY.

The URL-based ones work as they previously did, while the non-URL ones
are used for repository name-based lookups only.

This will require that people update their scripts or .reviewboardrc
files if passing the old values for the settings, but in the long-run,
it should avoid a lot of problems.