Allow for using repository names with Subversion repositories in 0.5.x.
Repository names for Subversion repositories broke Subversion. On
0.5.x, the name is stuffed into the repository_url setting, which the
Subversion code assumes means we're operating on some possibly remote
repository.

The proper fix to separate out repository names and URLs is coming in
0.6, but for now, we check the repository_url whenever we use it to see
whether it's a URL or a name.

This is intended to stay only on the 0.5.x branch, and will not merge to
master.