List git commits in reverse order.
When posting a branch with many commits, it's likely that the first one is the
most important one (with the best description), and the later ones are fixups.
This change alters extract_summary and extract_description to reflect that.
This also brings us into line with how the hg client works.

Testing done:
Created a branch with several commits and posted it. Saw that the summary was
that of the first commit, and that the commits were listed in chronological
rather than reverse chronological order.

Fixes bug 3163.

Reviewed at https://reviews.reviewboard.org/r/5347/