Clearcase often fails in multi-site mode because oids are not yet available on server
When people are submitting a review request, the oid of the old file version
and new file version are sent and used to retrieve the filename. However, the
new filename is usually not yet synced.

This change replaces revision /streamname/0 by its clearcase predecessor, i.e.
parentstream/N (N is origin revision of file in parent stream). This doesn't
completely fix the problem, but reduces the number of errors. A matching change
in reviewboard handles the rest of the errors.

Testing done:
Created a clearcase review request using --ucm-activity (which is a pending
change) in a fresh stream (containing /streamname/0 versions).

Fixes bug 3183.

Reviewed at https://reviews.reviewboard.org/r/5184/