Improve the resulting URL output from rbt post.
The new change to display the diff URL was a nice addition, but
displaying both URLs staggered on consecutive lines felt a little hard
to read.

This aligns the URLs to help with the readability. It also adds a blank
line between and after, which helps separate the results from each other
and the prompt (something that's bothered me for a while). This makes it
easier to click the links without clicking the wrong thing.