Don't use _convert_symbolic_revision with 'BASE'
In the case that the BASE revision is not present on the current working
directory (such as when the latest synced change is on a branch, and cwd is
trunk), trying to convert the BASE symbol to a numbered revision will fail.

This change just passes through the string 'BASE' as the base revision, which
should be both more efficient and cover all the possible situations.

Fixes bug 3250.