Don't include Error Stream output in the Diff
We ran into a problem where the diff is generated properly and git doesn't throw an error, but it prints warnings to stderr during the diff generation. The error stream content creates a corrupted diff, which causes breakage on the Review Board server. This patch makes it so that the diffs don't include the stderr stream content.

Sample Warning:
   warning: CRLF will be replaced by LF in pom.xml.
   The file will have its original line endings in your working directory.