Don't include error stream output in git diffs
We ran into a problem where the diff is generated properly and git doesn't
throw an error, but it prints warnings to stderr during the diff generation.
The error stream content is included in the diff and this creates a corrupted
diff, which causes breakage on the Review Board server. This patch makes it so
that the generated diffs don't include the stderr stream content.

Sample Warning:
   warning: CRLF will be replaced by LF in pom.xml.
   The file will have its original line endings in your working directory.

Testing Done:
We tested that the scenario described no longer occurs when the stderr is not
included in the diff.

Reviewed at https://reviews.reviewboard.org/r/5615/