Fix -I with svn repositories.
The line that added the files to the diff command line was indented four spaces
too many, which made it only work if the code happened to drop into that part
of the conditional.

Testing done:
Posted a change to an SVN working directory for just a single file.

Reviewed at https://reviews.reviewboard.org/r/5726/