Fix problems with "key=" lines in hgrc files.
The code that parses hgrc was assuming that all key=value lines had a non-empty
value. This failed if the config included lines that didn't.

Testing done:
Created an hgrc file with this configuration and ran rbtools commands. Saw that
the bug reproduced before the change and didn't after.

Reviewed at https://reviews.reviewboard.org/r/5748/