Fix posting of perforce paths.
The path behavior for rbt post was broken in a couple ways, probably for quite
some time. The first issue was that we had changed the `p4 change` usage to use
marshalled results, but the run_p4 method didn't respect the ignore_errors flag
when marshalling, and would call die(). Once I fixed this, the code that checks
the status of the changelist couldn't cope with error states. With these fixes,
_get_changelist_status will return None, which will cause parse_revision_spec
to correctly determine that the given path is not actually a revision.

Testing done:
Ran unit tests.

Fixes bug 3336.

Reviewed at https://reviews.reviewboard.org/r/5794/