Added Depends On field to rbt post
Added the ability to write to the Depends On field for a review through the rbt
post tool. Through the use of the flag --depends-on, one can specify review
numbers (comma separated) that the review depends on. The new flag has been
added to the rbt post manual.

Testing Done:
The manual was tested by using 'make html' to ensure that the formatting was
correct. For testing the flag, I created two fake reviews (id: 1 and 2) on my
local machine. Then I made, added, and committed a change to the reviewboard
repository. I ran the post command with the flag --depends-on 1 and checked to
ensure it added it in the web view. I then tested with --depends-on "1, 2" to
ensure it supports multiple revisions. Finally, I tested --depends-on "test" to
ensure it was checking arguments.

Bugs closed: 3074

Reviewed at https://reviews.reviewboard.org/r/5799/