Properly generate parent diffs in git with multiple remotes
When posting a revision range with Git, the git diff would try hard to
determine whether or not a parent diff was required by checking to see if the
first revision in the range was contained in the remote. This works fine for
basic setups where there's a single remote that's the same one that Review
Board talks to.

Unfortunately, if a user has multiple remotes, and pushes commits to a remote
which is different from the one Review Board uses, rbt would think that a
parent diff isn't required when it really is.

This change gets rid of the fancy testing, and just checks that the merge base
isn't the same as the parent revision. The worst case here is that we end up
generating a parent diff when it's not truly required, but it fixes the
problems with multiple remotes.

Testing Done:
1) Set up a series of patches that depended on one another. (Patch 2 depends
   on changes from patch 1)
2) Pushed the patches to a remote git repository
3) Ran 'rbt post patch1:patch2'

Previous behavior:
The file was not found in the repository (HTTP 400, API Error 207)

Behavior after this patch:
Review was properly uploaded.

Reviewed at https://reviews.reviewboard.org/r/5825/