Make 'rbt status' work better with public local sites.
With public local sites, it's possible that the 'user' link in the session
resource will be broken. This is because the users resource only includes users
who are part of the given local site.

This change makes 'rbt status' work a little better in this case. We don't
actually need any data from the user resource, since all we're looking up here
is the username. In this case, we can just use the title attribute of the link,
which also happens to save us an API request.

Testing done:
* Verified that 'rbt status' still works.
* Ran unit tests.

Fixes bug 3350.

Reviewed at https://reviews.reviewboard.org/r/5863/