Commit ID is not included in review request if --include is used on rbt post.
Previously when --include was used on an rbt post, it would take the commit ID
of the current commit. The issue is that if someone uploaded multiple files
from the same commit separately, the identical commit IDs would cause conflicts
in the ReviewRequests. The solution is to not include the commit ID when
"--include" is used on an rbt post.

Testing Done:
Made a commit with two files in the change. I was able to post and publish both
without error, which would not happen normally. When looking on the review
board web ui, there was no commit id associated with either.

Bugs closed: 3328

Reviewed at https://reviews.reviewboard.org/r/5818/