Fix error running 'rbt patch' with Mercurial repositories.
Running `rbt patch` with a Mercurial repository failed with the error "patch:
unrecognized option '--config'" because _execute() in
rbtools/clients/mercurial.py appends a --config option for loading hgrc files.
This option is used for hg commands, but is not a valid option for `patch`. To
fix this, we run `patch` by calling execute().

Testing Done:
- Ran unit tests.
- Ran `rbt patch` with a Mercurial repository, and saw that the patch was
  applied successfully.

Reviewed at https://reviews.reviewboard.org/r/5894/