Don't error out if receiving an unexpected line in 'svn status'.
When 'svn status', we'd attempt to look for particular data to indicate
whether there's any files scheduled to retain history in the commit. We
assumed there would be a 4th column to check, but this doesn't always
appear to be the case. We now catch any IndexErrors here and ignore
them.

Reviewed at https://reviews.reviewboard.org/r/5925/