Prioritize the repository name over URL when posting changes.
rbt post was assuming that the repository URL provided should take
precedence over the repository name. The problem is that Subversion can
be given a repository URL that it can use for communicating with the
repository, but that URL isn't necessarily the one we want for lookup.

The Subversion issue should be fixed separately, but it's safe to say
that the repository name is a better option if provided. We now use that
over the URL if we have it.

Fixes bug #3277.
Reviewed at https://reviews.reviewboard.org/r/5960/