Bring over a better UnitTest class from Review Board.
This class lets us have docstrings that wrap lines for the unit test
summaries, instead of constraining them to a single line.

This does somewhat duplicate the purpose of RBTestBase in
rbtools/utils/, but that should maybe be split up into a set of mixins
or a new class, as that's really more for client tests and not that
useful for API tests.

Reviewed at https://reviews.reviewboard.org/r/5962/