Fix post-commit handling of perforce 'branch' operations.
There was some code in here to handle 'branch' operations on files already, but
it wasn't in the list of whitelisted operations, and we'd therefore error out
with "Unsupported action type "branch" for <filename>".

This change reorders the renaming of actions with the whitelist, to simplify
the list a bit and fix things.

Testing done:
* Ran unit tests
* Submitted a change involving a branched file and saw that it got posted as a
  new file.

Fixes bug 3410.

Reviewed at https://reviews.reviewboard.org/r/5975/